feat(web): waitlist as root landing page
Replace the placeholder home page with the waitlist signup page.
Remove AppHeader from the layout (page owns its own HarmoniaBrandHeader),
drop the /waitlist route (now lives at /), and update page metadata.
