+++ apps/web/src/app/layout.tsx
@@ -1,12 +1,12 @@
 import "@harmonia/ui/styles/globals.css";
 import type { Metadata } from "next";
 
-import AppHeader from "@/components/app-header";
 import AppProviders from "@/components/app-providers";
 
 export const metadata: Metadata = {
-	title: "harmonia",
-	description: "harmonia",
+	title: "Harmonia — Your Spotify library, intelligently organized.",
+	description:
+		"Harmonia uses AI to analyze your music library and automatically create meaningful playlists from your songs.",
 };
 
 export default function RootLayout({
@@ -17,12 +17,7 @@ export default function RootLayout({
 	return (
 		<html lang="en" suppressHydrationWarning>
 			<body className="antialiased">
-				<AppProviders>
-					<div className="grid h-svh grid-rows-[auto_1fr]">
-						<AppHeader />
-						{children}
-					</div>
-				</AppProviders>
+				<AppProviders>{children}</AppProviders>
 			</body>
 		</html>
 	);
